@@ -1,4 +1,4 @@
-﻿import React, { useContext } from 'react';
+﻿import React from 'react';
 import { Accordion, Card } from 'react-bootstrap';
 import AppInsights from '../Services/AppInsights';
 
@@ -25,78 +25,64 @@ function Documentation() {
 
 
             <Accordion>
-
-
-                <Card>
-                    <Accordion.Toggle as={Card.Header} eventKey="0" className="faq-header" onClick={() => CardClick('0')}>
+                <Accordion.Item eventKey="0">
+                    <Accordion.Header className="faq-header" onClick={() => CardClick('0')}>
                         <span className="rh5">When will Vyntix be released?</span>
-                    </Accordion.Toggle>
-                    <Accordion.Collapse eventKey="0">
-                        <Card.Body>
+                    </Accordion.Header>
+                    <Accordion.Body>
                             <span className="rh6">
-                                Vyntix will be released in phases starting with the Vyntx FRED Client in late Q4 2020.  
-                                This will be followed by the Vyntix FRED Downloader, then the Vyntix Data Manager in the subsequent  two quarters.
-                                All components that are released ahead of Vyntix v1.0 are considered alpha and are subject to breaking changes unless stated otherwise.
+                                Vyntix components are being released on an ongoing basis.  
+                                Vyntix FRED client has been released. Desktop components are in development.  
+                                All components that are released ahead of Vyntix v1.0 are considered beta and are subject to breaking changes unless stated otherwise.
                             </span>
-                        </Card.Body>
-                    </Accordion.Collapse>
-                </Card>
+                    </Accordion.Body>
+                </Accordion.Item>
+
+                <Accordion.Item eventKey="1">
+                    <Accordion.Header  className="faq-header" onClick={() => CardClick('1')}>
+                            <span className="rh5">What will the monthly cost for a subscription be?</span>
+                    </Accordion.Header>
+                    <Accordion.Body>
+                                <span className="rh6">For the lowest tier (individuals and very small organizations) the monthly cost will be about the same as the cost for lunch for one person.  
+                                For larger organizations the cost will be about the same as bagels and coffee for the crew.
+                                Subscriptions for non-business users will always be free.</span>
+                    </Accordion.Body>
+                </Accordion.Item>
 
 
 
-                <Card>
-                    <Accordion.Toggle as={Card.Header} eventKey="1" className="faq-header" onClick={() => CardClick('1')}>
-                        <span className="rh5">What will the monthly cost for a subscription be?</span>
-                    </Accordion.Toggle>
-                    <Accordion.Collapse eventKey="1">
-                        <Card.Body>
-                            <span className="rh6">For the lowest tier (individuals and very small organizations) the monthly cost will be about the same as the cost for lunch for one person.  
-                            For larger organizations the cost will be about the same as bagels and coffee for the crew.
-                            Subscriptions for non-business users will always be free.</span>
-                        </Card.Body>
-                    </Accordion.Collapse>
-                </Card>
+                <Accordion.Item eventKey="2">
+                    <Accordion.Header className="faq-header" onClick={(e) => CardClick('2')}>
+                            <span className="rh5">Will Vyntix support data providers other than FRED?</span>
+                    </Accordion.Header>
+                    <Accordion.Body>
+                                <span className="rh6">Yes, we will support data providers as is feasable and as requested by our subscribers.
+                                Data providers receiving more requests will be prioritized.  Request your data provider <a href="https://github.com/leaderanalytics/Vyntix/issues" target="_blank">here</a>.
+                                </span>
+                    </Accordion.Body>
+                </Accordion.Item>
 
 
-
-                <Card>
-                    <Accordion.Toggle as={Card.Header} eventKey="2" className="faq-header" onClick={(e) => CardClick('2')}>
-                        <span className="rh5">Will Vyntix support data providers other than FRED?</span>
-                    </Accordion.Toggle>
-                    <Accordion.Collapse eventKey="2">
-                        <Card.Body>
-                            <span className="rh6">Yes, we will support data providers as is feasable and as requested by our subscribers.
-                            Data providers receiving more requests will be prioritized.  Request your data provider <a href="https://github.com/leaderanalytics/Vyntix/issues" target="_blank">here</a>.
-                            </span>
-                        </Card.Body>
-                    </Accordion.Collapse>
-                </Card>
+                <Accordion.Item eventKey="3">
+                    <Accordion.Header className="faq-header" onClick={() => CardClick('3')}>
+                            <span className="rh5">What are the differences beween the subscription plans?</span>
+                    </Accordion.Header>
+                    <Accordion.Body>
+                                <span className="rh6">There is only one subscription plan.  It is free for non-business use or priced differently based on revenue of the subscriber.</span>
+                    </Accordion.Body>
+                </Accordion.Item>
 
 
-                <Card>
-                    <Accordion.Toggle as={Card.Header} eventKey="3" className="faq-header" onClick={() => CardClick('3')}>
-                        <span className="rh5">What are the differences beween the subscription plans?</span>
-                    </Accordion.Toggle>
-                    <Accordion.Collapse eventKey="3">
-                        <Card.Body>
-                            <span className="rh6">There is only one subscription plan.  It is priced differently based on the company size of the subscriber.</span>
-                        </Card.Body>
-                    </Accordion.Collapse>
-                </Card>
-
-
-                <Card>
-                    <Accordion.Toggle as={Card.Header} eventKey="4" className="faq-header" onClick={() => CardClick('4')}>
-                        <span className="rh5">Does my Vyntix subscription include the cost of data?</span>
-                    </Accordion.Toggle>
-                    <Accordion.Collapse eventKey="4">
-                        <Card.Body>
-                            <span className="rh6">No.  At the time of this writng Vyntix only supports FRED - which is free to the general public.  In the future support may be added
-                            for data providers who charge for their data.  In these cases the subscriber must purchase data directly from the provider.
-                            </span>
-                        </Card.Body>
-                    </Accordion.Collapse>
-                </Card>
+                <Accordion.Item eventKey="4">
+                    <Accordion.Header className="faq-header" onClick={() => CardClick('4')}>
+                            <span className="rh5">Does my Vyntix subscription include the cost of data?</span>
+                    </Accordion.Header>
+                    <Accordion.Body>
+                                <span className="rh6">No.  At the time of this writng Vyntix only supports FRED - which is free to the general public.  In the future support may be added
+                                for data providers who charge for their data.  In these cases the subscriber must purchase data directly from the provider.
+                                </span>
+                    </Accordion.Body>
+                </Accordion.Item>
             </Accordion>
 
             <div>
@@ -116,8 +102,8 @@ function Documentation() {
                         </div>
 
                         <div className="section-body rh6 rp1"> 
-                            <p><b>Status:</b> In development</p>
-                            <p><b>Expected:</b> Q4 2020</p>
+                            <p><b>Status:</b> Released</p>
+                            <p><b>Release Date:</b> Q4 2022</p>
                             <p><b>Description:</b> Access the FRED API and creates data objects that are usable in code.</p>
                             <p><a href="/docs" target="_blank">Documentation</a></p>
                         </div>
@@ -125,13 +111,13 @@ function Documentation() {
 
                     <div className="rm2">
                         <div className="pageBanner section-header rp1">
-                            <span className="rh4">Vyntix FRED Downloader</span>
+                            <span className="rh4">Vyntix FRED Downloader (Observer)</span>
                         </div>
 
                         <div className="section-body rh6 rp1">
                             <p><b>Status:</b> In development</p>
-                            <p><b>Expected:</b> Q1 2021</p>
-                            <p><b>Description:</b> Cross platform (Linux/Mac/Windows) command line utilitiy that a downloads data from FRED and saves it to a file or a database.</p>
+                            <p><b>Expected:</b> Q4 2023</p>
+                            <p><b>Description:</b> Cross platform (Linux/Mac/Windows) command line utilitiy that a downloads data from FRED and saves it to a file or a database.  A desktop version is also in development.</p>
                             <p><a href="/docs" target="_blank">Documentation</a></p>
                         </div>
                     </div>
@@ -144,7 +130,7 @@ function Documentation() {
 
                         <div className="section-body rh6 rp1">
                             <p><b>Status:</b> In development</p>
-                            <p><b>Expected:</b> Q2 2021</p>
+                            <p><b>Expected:</b> Q4 2024</p>
                             <p><b>Description:</b> Windows desktop utility for managing data from multiple data providers.</p>
                             <p><a href="/docs" target="_blank">Documentation</a></p>
                         </div>
@@ -157,7 +143,7 @@ function Documentation() {
 
                         <div className="section-body rh6 rp1">
                             <p><b>Status:</b> In development</p>
-                            <p><b>Expected:</b> Q3 2021</p>
+                            <p><b>Expected:</b> Q3 2024</p>
                             <p><b>Description:</b> An API for creating and managing forecasts.</p>
                             <p><a href="/docs" target="_blank">Documentation</a></p>
                         </div>
@@ -171,7 +157,7 @@ function Documentation() {
 
                         <div className="section-body rh6 rp1">
                             <p><b>Status:</b> In development</p>
-                            <p><b>Expected:</b> Q3 2021</p>
+                            <p><b>Expected:</b> Q4 2024</p>
                             <p><b>Description:</b> Windows desktop application for managing forecasts.</p>
                             <p><a href="/docs" target="_blank">Documentation</a></p>
                         </div>
