@@ -6,13 +6,16 @@ import './App.scss';
 import './Assets/fonts/entsans.ttf';
 import { PublicClientApplication } from "@azure/msal-browser";
 import { msalConfig } from "./msalconfig";
+import { createRoot } from 'react-dom/client';
 
+const container = document.getElementById('root');
+const root = createRoot(container!);
 export const msalInstance = new PublicClientApplication(msalConfig);
-ReactDOM.render(
+
+root.render(
     <React.StrictMode>
         <App pca={msalInstance} />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
